Move classes to dedicated directories
