+++ app/Collections/RestrictedChannelsCollection.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App;
+namespace App\Collections;
 
+use App\Models\User;
 use Illuminate\Database\Eloquent\Collection as EloquentCollection;
 use Illuminate\Support\Facades\Auth;
 
@@ -11,7 +12,7 @@ class RestrictedChannelsCollection extends EloquentCollection
      * Return only channels that the user has a permission for.
      *
      * @param  string  $permission
-     * @param  \App\User|null  $user
+     * @param  \App\Models\User|null  $user
      * @return $this
      */
     public function withPermission(string $permission, User $user = null)
