+++ app/Models/Channel.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Traits\RestrictedChannels;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
