+++ app/Models/Course.php
@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Traits\HasLocation;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
