+++ app/Models/Occupation.php
@@ -3,6 +3,7 @@
 namespace App\Models;
 
 use App\Exceptions\UnknownOccupationStatusException;
+use App\Traits\HasLocation;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Log;
