+++ app/Models/Post.php
@@ -4,6 +4,9 @@ namespace App\Models;
 
 use App\Events\PostCreated;
 use App\Events\PostUpdated;
+use App\Traits\Favoritable;
+use App\Traits\MentionsUsers;
+use App\Traits\RecordsActivity;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
