+++ app/Models/Thread.php
@@ -3,6 +3,8 @@
 namespace App\Models;
 
 use App\Events\ThreadPublished;
+use App\Traits\MentionsUsers;
+use App\Traits\RecordsActivity;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
