+++ app/Providers/AuthServiceProvider.php
@@ -3,9 +3,9 @@
 namespace App\Providers;
 
 use App\Http\Middleware\Authenticate;
+use App\Models\Profile;
 use App\Models\User;
 use App\Policies\UserPolicy;
-use App\Profile;
 use Illuminate\Contracts\Auth\Middleware\AuthenticatesRequests;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Gate;
