+++ app/Traits/HasReputation.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Traits;
 
 trait HasReputation
 {
