+++ tests/Unit/HasLocationTest.php
@@ -3,7 +3,7 @@
 namespace Tests\Unit;
 
 use App\Models\Location;
-use App\Profile;
+use App\Models\Profile;
 use Illuminate\Support\Arr;
 use Tests\TestCase;
