+++ tests/Unit/ProfileTest.php
@@ -4,7 +4,7 @@ namespace Tests\Unit;
 
 use App\Models\Course;
 use App\Models\Occupation;
-use App\Profile;
+use App\Models\Profile;
 use Tests\TestCase;
 
 class ProfileTest extends TestCase
